whatever I did

+++ src/collision.cc
@@ -1,7 +1,9 @@
 #include "collision.h"
 #include "config.h"
+#include <raymath.h>
 
-bool check_cell_collision(Vector2 cell_position, Vector2 position) {
+bool check_cell_collision(Vector2 cell_position, Vector2 position)
+{
   Rectangle rect;
 
   rect.x = cell_position.x;
@@ -14,11 +16,22 @@ bool check_cell_collision(Vector2 cell_position, Vector2 position) {
 }
 
 int check_cells_collision(std::vector<Vector2> cell_positions,
-                          Vector2 position) {
+                          Vector2 position)
+{
 
-  for (int i = 0; i < cell_positions.size(); i++) {
+  for (size_t i = 0; i < cell_positions.size(); i++)
+  {
 
-    if (check_cell_collision(cell_positions[i], position)) {
+    // const float distance = Vector2Distance(position, cell_positions[i]);
+
+    // Skip if the distance is too far
+    // if (distance > CELL_SIZE + 20)
+    // {
+    //   continue;
+    // }
+
+    if (check_cell_collision(cell_positions[i], position))
+    {
       return i;
     }
   }
